na.rpart {rpart}	R Documentation

Handles Missing Values in an Rpart Object

Description

Handles missing values in an "rpart" object.

Usage


na.rpart(x)

Arguments

x
a model frame.

Details

Default function that handles missing values when calling the function rpart.

It omits cases where part of the response is missing or all the explanatory variables are missing.

[Package rpart version 4.1-13 Index]