rpart.exp {rpart}	R Documentation

Initialization function for exponential fitting

Description

This function does the initialization step for rpart, when the response is a survival object. It rescales the data so as to have an exponential baseline hazard and then uses Poisson methods. This function would rarely if ever be called directly by a user.

Usage


rpart.exp(y, offset, parms, wt)

Arguments

y
the response, which will be of class Surv

offset
optional offset

parms
parameters controlling the fit. This is a list with components shrink and method. The first is the prior for the coefficient of variation of the predictions. The second is either "deviance" or "sqrt" and is the measure used for cross-validation. If values are missing the defaults are used, which are "deviance" for the method, and a shrinkage of 1.0 for the deviance method and 0 for the square root.

wt
case weights, if present

Value

a list with the necessary initialization components

Author(s)

Terry Therneau

See Also

rpart

[Package rpart version 4.1-13 Index]