rpconvert {rpart}	R Documentation

Update an rpart object

Description

Rpart objects changed (slightly) in their internal format in order to accommodate the changes for user-written split functions. This routine updates an old object to the new format.

Usage

rpconvert(x)

Arguments

x 	 an rpart object

Value

an updated object

See Also

rpart

[Package rpart version 3.1-45 Index]