Stage C Prostate Cancer

A set of 146 patients with stage C prostate cancer, from a study exploring the prognostic value of flow cytometry.

Usage

data(stagec)

Format

A data frame with 146 observations on the following 8 variables.

pgtime

Time to progression or last follow-up (years)

pgstat

1 = progression observed, 0 = censored

age

age in years

eet

early endocrine therapy, 1 = no, 2 = yes

g2

percent of cells in G2 phase, as found by flow cytometry

grade

grade of the tumor, Farrow system

gleason

grade of the tumor, Gleason system

ploidy

the ploidy status of the tumor, from flow cytometry. Values are diploid, tetraploid, and aneuploid

Details

A tumor is called diploid (normal complement of dividing cells) if the fraction of cells in G2 phase was determined to be 13% or less. Aneuploid cells have a measurable fraction with a chromosome count that is neither 24 nor 48, for these the G2 percent is difficult or impossible to measure.

Examples

require(survival)
#> Loading required package: survival
rpart(Surv(pgtime, pgstat) ~ ., stagec)
#> n= 146 
#> 
#> node), split, n, deviance, yval
#>       * denotes terminal node
#> 
#>  1) root 146 192.111100 1.0000000  
#>    2) grade< 2.5 61  44.799010 0.3634439  
#>      4) g2< 11.36 33   9.117405 0.1229835 *
#>      5) g2>=11.36 28  27.602190 0.7345610  
#>       10) gleason< 5.5 20  14.297110 0.5304115 *
#>       11) gleason>=5.5 8  11.094650 1.3069940 *
#>    3) grade>=2.5 85 122.441500 1.6148600  
#>      6) age>=56.5 75 103.062900 1.4255040  
#>       12) gleason< 7.5 50  66.119800 1.1407320  
#>         24) g2< 13.475 24  27.197170 0.8007306 *
#>         25) g2>=13.475 26  36.790960 1.4570210  
#>           50) g2>=17.915 15  20.332740 0.9789825 *
#>           51) g2< 17.915 11  13.459010 2.1714480 *
#>       13) gleason>=7.5 25  33.487250 2.0307290  
#>         26) g2>=15.29 10  11.588480 1.2156230 *
#>         27) g2< 15.29 15  18.939150 2.7053610 *
#>      7) age< 56.5 10  13.769010 3.1822320 *