summary.rpart {rpart}	R Documentation

Summarize a Fitted Rpart Object

Description

Returns a detailed listing of a fitted rpart object.

Usage


## S3 method for class 'rpart'
summary(object, cp = 0, digits = getOption("digits"), file, ...)

Arguments

object
fitted model object of class "rpart". This is assumed to be the result of some function that produces an object with the same named components as that returned by the rpart function.

digits
Number of significant digits to be used in the result.

cp
trim nodes with a complexity of less than cp from the listing.

file
write the output to a given file name. (Full listings of a tree are often quite long).

...
arguments to be passed to or from other methods.

Details

This function is a method for the generic function summary for class "rpart". It can be invoked by calling summary for an object of the appropriate class, or directly by calling summary.rpart regardless of the class of the object.

It prints the call, the table shown by printcp, the variable importance (summing to 100) and details for each node (the details depending on the type of tree).

See Also

summary, rpart.object, printcp.

Examples


## a regression tree
z.auto <- rpart(Mileage ~ Weight, car.test.frame)
summary(z.auto)

## a classification tree with multiple variables and surrogate splits.
summary(rpart(Kyphosis ~ Age + Number + Start, data = kyphosis))

[Package rpart version 4.1-13 Index]