text.rpart {rpart}	R Documentation

Place Text on a Dendrogram Plot

Description

Labels the current plot of the tree dendrogram with text.

Usage


## S3 method for class 'rpart'
text(x, splits = TRUE, label, FUN = text, all = FALSE,
     pretty = NULL, digits = getOption("digits") - 3, use.n = FALSE,
     fancy = FALSE, fwidth = 0.8, fheight = 0.8, bg = par("bg"),
     minlength = 1L, ...)

Arguments

x
fitted model object of class "rpart". This is assumed to be the result of some function that produces an object with the same named components as that returned by the rpart function.

splits
logical flag. If TRUE (default), then the splits in the tree are labeled with the criterion for the split.

label
For compatibility with rpart2, ignored in this version (with a warning).

FUN
the name of a labeling function, e.g. text.

all
Logical. If TRUE, all nodes are labeled, otherwise just terminal nodes.

minlength
the length to use for factor labels. A value of 1 causes them to be printed as ‘a’, ‘b’, ..... Larger values use abbreviations of the label names. See the labels.rpart function for details.

pretty
an alternative to the minlength argument, see labels.rpart.

digits
number of significant digits to include in numerical labels.

use.n
Logical. If TRUE, adds to label (\#events level1/ \#events level2/etc. for class, n for anova, and \#events/n for poisson and exp).

fancy
Logical. If TRUE, nodes are represented by ellipses (interior nodes) and rectangles (leaves) and labeled by yval. The edges connecting the nodes are labeled by left and right splits.

fwidth
Relates to option fancy and the width of the ellipses and rectangles. If fwidth < 1 then it is a scaling factor (default = 0.8). If fwidth > 1 then it represents the number of character widths (for current graphical device) to use.

fheight
Relates to option fancy and the height of the ellipses and rectangles. If fheight <1 then it is a scaling factor (default = 0.8). If fheight > 1 then it represents the number of character heights (for current graphical device) to use.

bg
The color used to paint the background to annotations if fancy = TRUE.

...
Graphical parameters may also be supplied as arguments to this function (see par). As labels often extend outside the plot region it can be helpful to specify xpd = TRUE.

Side Effects

the current plot of a tree dendrogram is labeled.

See Also

text, plot.rpart, rpart, labels.rpart, abbreviate

Examples


freen.tr <- rpart(y ~ ., freeny)
par(xpd = TRUE)
plot(freen.tr)
text(freen.tr, use.n = TRUE, all = TRUE)

[Package rpart version 4.1-13 Index]