xpred.rpart {rpart}	R Documentation

Return Cross-Validated Predictions

Description

Gives the predicted values for an rpart fit, under cross validation, for a set of complexity parameter values.

Usage

xpred.rpart(fit, xval=10, cp)

Arguments

fit 	 a rpart object.

xval 	 number of cross-validation groups. This may also be an explicit list of integers that define the cross-validation groups.

cp 	 the desired list of complexity values. By default it is taken from the cptable component of the fit.

Details

Complexity penalties are actually ranges, not values. If the cp values found in the table were .36, .28, and .13, for instance, this means that the first row of the table holds for all complexity penalties in the range [.36, 1], the second row for cp in the range [.28, .36) and the third row for [.13,.28). By default, the geometric mean of each interval is used for cross validation.

Value

a matrix with one row for each observation and one column for each complexity value.

See Also

rpart

Examples

fit <- rpart(Mileage ~ Weight, car.test.frame)
xmat <- xpred.rpart(fit)
xerr <- (xmat - car.test.frame$Mileage)^2
apply(xerr, 2, sum)   # cross-validated error estimate

# approx same result as rel. error from printcp(fit)
apply(xerr, 2, sum)/var(car.test.frame$Mileage) 
printcp(fit)

[Package rpart version 3.1-45 Index]